+++ variables.tf
@@ -78,7 +78,6 @@ variable "virtual_machine_config" {
     os_disk_write_accelerator_enabled = optional(bool, false)
     admin_username                    = optional(string, "loc_sysadmin")
     timezone                          = optional(string, "UTC")
-    zone                              = optional(string)
     availability_set_id               = optional(string)
     proximity_placement_group_id      = optional(string)
     patch_assessment_mode             = optional(string, "AutomaticByPlatform")
@@ -97,10 +96,7 @@ variable "virtual_machine_config" {
     condition     = (contains(["Premium_LRS", "Premium_ZRS"], var.virtual_machine_config.os_disk_storage_type) && var.virtual_machine_config.os_disk_write_accelerator_enabled == true  && var.virtual_machine_config.os_disk_caching == "None") || (var.virtual_machine_config.os_disk_write_accelerator_enabled == false)
     error_message = "os_disk_write_accelerator_enabled, can only be activated on Premium disks and caching deactivated."
   }
-  validation {
-    condition     = var.virtual_machine_config.zone == null || var.virtual_machine_config.zone == "1" || var.virtual_machine_config.zone == "2" || var.virtual_machine_config.zone == "3"
-    error_message = "Zone, can only be empty, 1, 2 or 3."
-  }
+
   description = <<-DOC
   ```
     hostname: Name of the host system.
@@ -117,7 +113,6 @@ variable "virtual_machine_config" {
       be activated on Premium_LRS disks and caching deactivated. Defaults to false.
     timezone: Optionally change the timezone of the VM. Defaults to UTC.
       (More timezone names: https://jackstromberg.com/2017/01/list-of-time-zones-consumed-by-azure/).
-    zone: Optionally specify an availibility zone for the vm.
     availability_set_id: Optionally specify an availibilty set for the vm.
     proximity_placement_group_id: (Optional) The ID of the Proximity Placement Group which the Virtual Machine should be assigned to.
     patch_assessment_mode: Specifies the mode of VM Guest Patching for the Virtual Machine.
@@ -150,6 +145,7 @@ variable "data_disks" {
   type = map(object({
     lun                        = number
     disk_size_gb               = number
+    zone                       = optional(string)
     caching                    = optional(string, "ReadWrite")
     create_option              = optional(string, "Empty")
     source_resource_id         = optional(string)
@@ -162,12 +158,22 @@ variable "data_disks" {
     error_message = "One or more of the lun parameters in the map are duplicates."
   }
   validation {
-    condition     = alltrue([for o in var.data_disks : contains(["Standard_LRS", "StandardSSD_LRS", "Premium_LRS", "StandardSSD_ZRS", "Premium_ZRS"], o.storage_account_type)])
-    error_message = "Possible values are Standard_LRS, StandardSSD_LRS, Premium_LRS, StandardSSD_ZRS and Premium_ZRS for storage_account_type"
+    condition     = alltrue([for o in var.data_disks : contains(["Standard_LRS", "StandardSSD_LRS", "Premium_LRS", "StandardSSD_ZRS", "Premium_ZRS", "PremiumV2_LRS"], o.storage_account_type)])
+    error_message = "Possible values are Standard_LRS, StandardSSD_LRS, Premium_LRS, StandardSSD_ZRS, Premium_ZRS and PremiumV2_LRS for storage_account_type"
+  }
+
+    validation {
+    condition = alltrue([
+      for disk in var.data_disks : (
+        disk.on_demand_bursting_enabled == false ||
+        disk.disk_size_gb > 512
+      )
+    ])
+    error_message = "on_demand_bursting_enabled` can only be set to true when `disk_size_gb` is larger than 512GB."
   }
   validation {
-    condition = alltrue([for o in var.data_disks : (
-      (o.write_accelerator_enabled == true && contains(["Premium_LRS", "Premium_ZRS"], o.storage_account_type) && contains(["None", "ReadOnly"], o.caching)) ||
+    condition = alltrue([for o in var.data_disks: (
+      (o.write_accelerator_enabled == true && contains(["Premium_LRS", "Premium_ZRS"], o.storage_account_type) && contains(["None"], o.caching)) ||
       (o.write_accelerator_enabled == false)
     )])
     error_message = "write_accelerator_enabled, can only be activated on Premium disks and caching deactivated."
@@ -177,7 +183,44 @@ variable "data_disks" {
       (o.on_demand_bursting_enabled == true && contains(["Premium_LRS", "Premium_ZRS"], o.storage_account_type)) ||
       (o.on_demand_bursting_enabled == false)
     )])
-    error_message = "If enable on demand bursting, possible storage_account_type values are Premium_LRS and Premium_ZRS."
+    error_message = "If enable on demand bursting, possible storage_account_type values are Premium_LRS, Premium_ZRS"
+  }
+  validation {
+    condition = alltrue([
+      for v in var.data_disks :
+      (
+        v.zone == null ? true : contains(["1", "2", "3"], v.zone)
+      )
+    ])
+    error_message = "Zone must be null, '1', '2' or '3'."
+  }
+  validation {
+    condition = alltrue([
+      for v in var.data_disks :
+      (
+        (v.storage_account_type != "PremiumV2_LRS") || 
+        (v.zone != null)
+      )
+    ])
+    error_message = "PremiumV2_LRS storage_account_type requires zone to be set. Please set zone to 1, 2 or 3."
+  }
+  validation {
+    condition = (
+      length(distinct([
+        for v in var.data_disks : v.zone
+      ])) <= 1
+    )
+    error_message = "All disks must be in the same zone or zone must be empty."
+  }
+  validation {
+    condition = alltrue([
+      for v in var.data_disks :
+      (
+        (v.storage_account_type != "PremiumV2_LRS") ||
+        (v.caching == "None")
+      )
+    ])
+    error_message = "When storage_account_type is 'PremiumV2_LRS', caching must be set to 'None'."
   }
   validation {
     condition     = alltrue([for k, v in var.data_disks : !strcontains(k, "-")])